Move homepage, fix tests
- follow the structure we follow with the other components, css
