+++ src/assets/Homepage.css
@@ -2,7 +2,7 @@
   margin: 0px 0px -15px 0px;
   padding: 0;
   background-color: rgba(238, 238, 238, 0.4) !important;
-  background: url("../../images/landing-page-background.png");
+  background: url("../images/landing-page-background.png");
   background-blend-mode: color;
   background-size: cover;
   background-position: center;
