+++ src/components/App.js
@@ -4,7 +4,7 @@ import UserProfile from '../containers/UserProfile'
 import { Route, Switch } from 'react-router-dom'
 import ProjectsList from '../containers/ProjectsList'
 import ProjectInfo from '../containers/ProjectInfo'
-import Homepage from '../components/homepage/Homepage'
+import Homepage from '../components/Homepage'
 import Footer from '../components/Footer'
 import Navbar from './navbar/Navbar'
 import About from '../containers/About'
